refactor(skills): 抽取 InstalledSkillCard 为独立组件
