+++ src/renderer/pages/settings/skills/index.tsx
@@ -5,8 +5,8 @@
  */
 
 import React, { useCallback, useEffect, useRef, useState } from 'react';
-import { Button, Spin, Message, Input, Progress, Modal, Popconfirm, Switch, Tooltip } from '@arco-design/web-react';
-import { Search, Delete, Close, Shield, Lightning, UploadOne, Install, Share, Plus, Check } from '@icon-park/react';
+import { Button, Spin, Message, Input, Progress, Modal, Popconfirm, Tooltip } from '@arco-design/web-react';
+import { Search, Delete, Close, Shield, Lightning, UploadOne, Share, Plus, Check } from '@icon-park/react';
 import { IconDownload, IconRefresh } from '@arco-design/web-react/icon';
 import { PackagePlus } from 'lucide-react';
 import classNames from 'classnames';
@@ -26,6 +26,7 @@ import { useAuth } from '@/renderer/context/AuthContext';
 import { useAppMode } from '@/renderer/hooks/useAppMode';
 import { SkillAuditSummary, SkillAuditDetailModal, SkillAuditReportModal } from '@/renderer/components/SettingsModal/contents/SkillAuditReport';
 import SkillCard from './components/SkillCard';
+import InstalledSkillCard from './components/InstalledSkillCard';
 
 // ==================== Helpers ====================
 
@@ -167,101 +168,6 @@ export function getInstalledSkillBadgeCount(installedList: IInstalledSkillInfo[]
 }
 
 // ==================== InstalledSkillCard Component ====================
-
-const InstalledSkillCard: React.FC<{
-  skill: IInstalledSkillInfo;
-  onUninstall: () => void;
-  uninstalling: boolean;
-  onToggleEnabled?: (enabled: boolean) => void;
-  togglingEnabled: boolean;
-  onClick?: () => void;
-  hasUpdate?: boolean;
-  onUpdate?: () => void;
-  updating?: boolean;
-  /** Enterprise mode: publish button element */
-  enterprisePublishButton?: React.ReactNode;
-  /** Enterprise mode: whether to hide uninstall button (only custom skills can be uninstalled) */
-  hideUninstall?: boolean;
-}> = ({ skill, onUninstall, uninstalling, onToggleEnabled, togglingEnabled, onClick, hasUpdate, onUpdate, updating, enterprisePublishButton, hideUninstall }) => {
-  const { displayName, description, icon, emoji } = getInstalledSkillDisplay(skill);
-  const displayVersion = normalizeSkillVersion(skill.version);
-  const canUninstall = !skill.isBuiltin && !hideUninstall;
-  const canToggleEnabled = !!skill.meta && !skill.isBuiltin;
-  const hasDetail = !!skill.meta;
-  const isEnabled = skill.enabled;
-  const { t } = useTranslation();
-
-  return (
-    <div className={classNames('item-card group flex items-start gap-3 relative overflow-hidden', !isEnabled && 'opacity-65', !hasDetail && 'cursor-default')} onClick={hasDetail ? onClick : undefined}>
-      {/* Icon */}
-      <div className='w-48px flex-shrink-0'>
-        <div className='w-48px h-48px rd-8px overflow-hidden bg-fill-2'>
-          {icon ? (
-            <img src={icon} alt={displayName} className='w-full h-full object-cover' onError={handleSkillIconError} />
-          ) : emoji ? (
-            <div className='w-full h-full f-center text-22px'>{emoji}</div>
-          ) : (
-            <div className='w-full h-full f-center bg-primary-light'>
-              <Lightning size='22' className='text-primary' />
-            </div>
-          )}
-        </div>
-      </div>
-
-      {/* Content */}
-      <div className='flex-1 min-w-0'>
-        <div className='flex items-center gap-6px pr-58px min-w-0'>
-          <span className='flex-1 min-w-0 font-medium text-13px text-foreground truncate'>{displayName}</span>
-          {!skill.isBuiltin && displayVersion && <span className='px-5px py-0px bg-fill-3 text-secondary text-10px rd-3px whitespace-nowrap flex-shrink-0 leading-18px'>v{displayVersion}</span>}
-        </div>
-        <div className='mt-3px min-h-30px'>{description ? <div className='text-11px text-secondary line-clamp-2 leading-15px'>{description}</div> : <div className='text-11px text-tertiary italic line-clamp-2 leading-15px'>{skill.name}</div>}</div>
-      </div>
-
-      {/* Actions - top right */}
-      <div className='absolute top-10px right-10px flex items-center gap-6px' onClick={(e) => e.stopPropagation()}>
-        {hasUpdate && (
-          <Tooltip content={t('settings.skill.updateAvailable', { defaultValue: '可更新' })}>
-            <button
-              type='button'
-              className='store-action-icon'
-              onClick={(e) => {
-                e.stopPropagation();
-                onUpdate?.();
-              }}
-            >
-              {updating ? <Spin size={10} /> : <Install size='13' />}
-            </button>
-          </Tooltip>
-        )}
-        {enterprisePublishButton}
-        {canToggleEnabled && <Switch size='small' checked={isEnabled} loading={togglingEnabled} onChange={(checked) => onToggleEnabled?.(checked)} className={isEnabled ? '!bg-[var(--ui-accent-orange)] !border-[var(--ui-accent-orange)]' : ''} />}
-        {skill.isBuiltin ? (
-          <Tooltip content='内置技能'>
-            <div className='store-action-icon cursor-default'>
-              <Shield size='15' />
-            </div>
-          </Tooltip>
-        ) : !canUninstall ? (
-          <Tooltip content='内置技能无法卸载'>
-            <div className='store-action-icon cursor-default opacity-30'>
-              <Shield size='14' />
-            </div>
-          </Tooltip>
-        ) : uninstalling ? (
-          <Spin size={14} />
-        ) : (
-          <Popconfirm title='确认卸载该技能？' onOk={onUninstall} okText='卸载' cancelText='取消' okButtonProps={{ status: 'danger' }}>
-            <Tooltip content={t('common.delete', { defaultValue: '删除' })}>
-              <div className='store-action-icon store-action-icon--danger'>
-                <Delete size='15' />
-              </div>
-            </Tooltip>
-          </Popconfirm>
-        )}
-      </div>
-    </div>
-  );
-};
 
 // ==================== SkillDetailModal Component ====================
